Stream R: 30 P1 items — grad-flow, Float64 precision, parity
audit-3 P1.  Lands every Stream R item (R.1–R.30).

Float64 precision (R.1, R.2, R.6):
* autograd::clamp / pow public APIs widened float → double.
  ClampBackward min_/max_ fields and saved_attributes scalars now
  double.  Tensor pow(Variable, double) binding stops dropping
  precision via static_cast<float>.
* JVP activation/SDPA scale/p/beta/alpha narrowing removed at
  10 sites in jvp_rules.cpp; helper sigs (jvp_clamp / jvp_leaky_
  relu / jvp_elu / jvp_softplus / jvp_norm / jvp_sdpa_forward)
  widened double.
* composed_attention_backward_variable scale: float → double;
  FlashAttentionBackward / FusedAttentionBackward /
  FlexAttentionBackward scale_ fields widened.

var / std / IRFFT (R.3, R.4, R.7, R.8):
* var / std autograd backward now branches on unbiased
  (default true).  VarBackward / StdBackward gained unbiased_
  field + 3-arg ctor + saved_attributes(Unbiased).
* vmap dim_shifted_passthrough(AttrKey::Dim) now registered for
  Softmax / LogSoftmax / CumSum / CumProd.  Gather/IndexSelect/
  Scatter/ScatterAdd use a new multi_input_dim_loop_rule that
  delegates to vmap_loop_and_stack (their captured index/src
  Variables aren't reachable from saved_attributes).
* IRFFTBackward saves n_orig_ on forward, passes it to rfft
  instead of std::nullopt (zero-padded / truncated transforms
  now produce the right bin count).

21 severed-grad sites (R.5, R.9):
* Linear ops (ConjBackward / RealBackward / ImagBackward /
  DeviceTransferBackward / SparseToDenseBackward /
  DenseToSparseBackward) gained is_higher_order_stub() override
  — second derivative is structurally zero, so the engine's
  stub counter now reports honestly.
* Non-linear ops rewritten as Variable-level composition with
  live grad_fn: SliceBackward, SpMMBackward, SpMVBackward,
  SpGEMMBackward, SparseTriSolveBackward, SparseSoftmaxBackward,
  SparseLogSoftmaxBackward, CholeskyInverseBackward,
  LinalgLDLSolveBackward, TensorInvBackward, TensorSolveBackward,
  CosineSimilarityBackward, RenormBackward.  RenormBackward also
  picks up the previously-dropped correction term.
* Honest stubs (is_higher_order_stub() = true + WARN_ONCE):
  IndexBackward, GridSampleBackward, AffineGridBackward — first-
  order opaque kernels with no Variable-level decomposition.

Backend hygiene (R.10–R.14):
* ROCm scatter_hip both overloads now throw on dim out of range
  (matching gather_hip).
* ROCm BF16 atomicAddHelper does explicit RNE rounding via
  uint32 lsb+0x7fff bias (was value-truncating ctor).
* CUDA cudaMemsetAsync TENZOR_CUDA_CHECK sweep — 33 sites across
  batchnorm.cu, advanced.cu, pooling.cu, indexing.cu, cudnn_ops.
  cu, math.cu, transform.cu, activations.cu, cuda_kernel_registry.
  cpp.
* Vulkan FP64 host-side gate: new vulkan::ensure_fp64_supported()
  helper called from dispatchFull/Ones/Rand/Randn/MaskedSelect/
  Where/Interpolate/ROIAlign{Forward,Backward}/ArgSort/Cast.
* OneAPI conv3d / conv3d_backward_input / conv3d_backward_weight
  wrap pd + dnnl_stream.wait() in catch(const dnnl::error&) +
  TENZOR_LOG_ERROR + rethrow.

nn/optim/distributed (R.15–R.21):
* Module::to(DType) appends train(training_) so sub-modules
  recreated during dtype move keep parent's mode.
* Adam / AdamW / NAdam / RAdam / SparseAdam / LAMB allocate
  state buffers in Float32 when params are F16/BF16; step
  upcasts grad+param to F32 for math, casts back on assignment.
  (PyTorch master-weights AMP convention.)
* InstanceNorm1d / InstanceNorm2d push bias into input_vars when
  affine — backward bias gradient was always zero.
* FSDP / FSDP2 param-replace: in-place zero_+add_ (the codebase's
  public copy_ surrogate) preserves Storage handle.  Saved
  activations no longer see stale views during shard↔full
  transitions.
* Embedding GPU backward populates sparse_grad via accumulate_
  sparse_grad (mirrors the CPU path).  SparseAdam reads
  param.sparse_grad() first; falls back to dense rows-mask.
* Attention normalises attn_mask to 4D up front so 2D [Sq,Sk]
  and 3D [N*H,Sq,Sk] correctly broadcast against
  key_padding_mask's [N,1,1,Sk].

Python frontend (R.22–R.30):
* autograd.grad / func.grad / vmap / jacrev / jacfwd / hessian /
  jvp / hvp / vhp / vjp / gradcheck / gradgradcheck wrap their
  outer body in py::gil_scoped_release; the inner gil_scoped_
  acquire inside each cpp_fn re-acquires for Python callbacks.
* PyAutocastContext / PyNoGradContext / PyEnableGradContext
  store vector<unique_ptr<Guard>> instead of single pointer so
  with-statement re-entry stacks correctly.
* default_collate handles dataclass and NamedTuple paths
  ahead of the generic tuple/list fallback.
* _pin_memory_batch removes the hasattr(_core, 'pin_memory')
  silent fallback — call unconditionally so the M.8 contract
  is enforced inside worker subprocesses too.
* numpy_interop FP8 branch points at .to(Float32).numpy() (not
  .dequantize() which doesn't exist on FP8).  numpy_dtype_to_
  tenzor gained ml_dtypes.float8_e4m3fn / float8_e5m2 detection.
* functional.pyi aliases `from tenzor import Variable as Tensor`
  so existing Tensor-typed signatures resolve to Variable.
* gelu accepts approximate (forwards to existing binding);
  softplus binding extended with threshold parameter (PyTorch
  semantics via Variable where).
* New python/tenzor/fft.pyi enumerates 21 fft functions.
* Variable index branches added to is_fancy_element + bare-key
  path + per-tuple-element loop in __getitem__; unwraps via
  .tensor() and validates Int32/Int64.

Verified: full ninja build (1898 targets) clean in single -j1
shell.  No tests run (changes are mostly internal correctness
fixes; ctest patterns will run at end-of-stream gate per the
audit-3 plan).

+++ include/tenzor/autograd/jvp_rules.hpp
@@ -39,10 +39,10 @@ auto jvp_relu(const DualTensor& x) -> DualTensor;
 auto jvp_sigmoid(const DualTensor& x) -> DualTensor;
 auto jvp_tanh(const DualTensor& x) -> DualTensor;
 auto jvp_gelu(const DualTensor& x) -> DualTensor;
-auto jvp_leaky_relu(const DualTensor& x, float negative_slope = 0.01f) -> DualTensor;
-auto jvp_elu(const DualTensor& x, float alpha = 1.0f) -> DualTensor;
+auto jvp_leaky_relu(const DualTensor& x, double negative_slope = 0.01) -> DualTensor;
+auto jvp_elu(const DualTensor& x, double alpha = 1.0) -> DualTensor;
 auto jvp_selu(const DualTensor& x) -> DualTensor;
-auto jvp_softplus(const DualTensor& x, float beta = 1.0f) -> DualTensor;
+auto jvp_softplus(const DualTensor& x, double beta = 1.0) -> DualTensor;
 auto jvp_mish(const DualTensor& x) -> DualTensor;
 auto jvp_hardswish(const DualTensor& x) -> DualTensor;
 /// @}
@@ -311,7 +311,7 @@ auto jvp_adaptive_avgpool_3d(const DualTensor& x, int64_t output_d, int64_t outp
 auto jvp_sdpa_forward(const DualTensor& Q,
                       const DualTensor& K,
                       const DualTensor& V,
-                      float scale,
+                      double scale,
                       bool causal) -> DualTensor;
 /// @}
